+++ src/main/java/origami_editor/editor/Editor.java
@@ -30,7 +30,7 @@ public class Editor {
     }
 
     private void createUIComponents() {
-     }
+    }
 
     /**
      * Method generated by IntelliJ IDEA GUI Designer
